% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collar_history.R
\name{collar_history}
\alias{collar_history}
\title{Conversion of adult capture data into capture history data}
\usage{
collar_history(capdat, df)
}
\arguments{
\item{capdat}{data.frame with all the capture information}

\item{df}{format of CaptureDate and MortalityDate}
}
\value{
Returns a data.frame with start and end date of every individual serial number along with an assigned animal ID)
}
\description{
Convert capture data into a record of every serial number and start and end dates of every individual captured
}
\examples{
\donttest{Hist<-collar_history(capdat = data.frame(capdat), df = '\%m/\%d/\%Y')}
}
\keyword{ID,}
\keyword{animal}
\keyword{capture,}
\keyword{history}
