% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Creation of Common Raster Grid}
\usage{
create_grid(xmin, xmax, ymin, ymax, res, projstring)
}
\arguments{
\item{xmin}{min x coordinate}

\item{xmax}{max x coordinate}

\item{ymin}{min y coordinate}

\item{ymax}{max y coordinate}

\item{res}{cell size resolution}

\item{projstring}{proj4string for raster}
}
\value{
Returns a raster for use in other functions
}
\description{
Create a common raster grid for prediction
}
\examples{
\donttest{hr<-create_grid(xmin = 620000, xmax = 702000, ymin = 4510000, ymax = 4620000, res = 10, projstring = '+proj=utm +zone=12 +ellps=GRS80 +datum=NAD83 +units=m +no_defs')}
}
\keyword{ID,}
\keyword{animal}
\keyword{capture,}
\keyword{homerange}
