% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_HR.R
\name{create_HR}
\alias{create_HR}
\title{Create homerange for an individual}
\usage{
create_HR(dat, grid, aidname)
}
\arguments{
\item{dat}{data.frame with relocation data}

\item{grid}{grid as output by DEERP::create_grid}

\item{aidname}{name of aid column}
}
\value{
Returns a raster HR for an individual on a common grid
}
\description{
Create a kernel UD HR for an individual using a custom grid
}
\examples{
\donttest{hr<-HRFun(dat = dat, grid = grid)}
}
\keyword{ID,}
\keyword{animal}
\keyword{capture,}
\keyword{homerange}
