% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ssf.R
\name{fit_ssf}
\alias{fit_ssf}
\title{Fit an SSF Model}
\usage{
fit_ssf(dat, covars, lcname = "LCClass_03272019", rfnum = NULL)
}
\arguments{
\item{dat}{data.frame from ssf_sample}

\item{covars}{character vector of covars for the model}

\item{lcname}{name of land cover covariate}

\item{rfnum}{if not null, the column name of the rf rsf layer}
}
\value{
Returns SSF model object
}
\description{
Construct an SSF model from data created with ssf_sample
}
\examples{
\donttest{system.time({ mods<- parallel::mclapply(dd,fit_ssf, covars=c('sl_','ta_','DistToMainStream','DistPrimRoad','DistSecRoad','Elevation','Slope','Aspect','TPI','TRI','Roughness','ClosestDist'),lcname='LCClass_03272019',mc.cores=18) })}
amt::strata(AID)')
amt::strata(AID)') 
}
\keyword{competition}
\keyword{direct}
\keyword{ssf,}
