% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_comp.R
\name{nearest_comp}
\alias{nearest_comp}
\title{Compute distance to nearest potential competitor}
\usage{
nearest_comp(x, fd, td = 0.5, ncpu = 84)
}
\arguments{
\item{x}{data.frame to calculate distance to}

\item{fd}{entire dataset that includes animals to calculate distance to}

\item{td}{difference in time to look for animals, in hours}

\item{ncpu}{number of cpus to use for calculations}
}
\value{
SSF Model ready data with distance to nearest competitor
}
\description{
Determine the distance to nearest competitor within given timeframe
}
\examples{
\donttest{system.time({ od<-lapply(dd,nearest_comp, fd=df, td=0.5, ncpu=84) })}
}
\keyword{coefficients,}
\keyword{competition,}
\keyword{direct}
\keyword{plots}
\keyword{ssf,}
