% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_coefs.R
\name{parse_coefs}
\alias{parse_coefs}
\title{Parse coefficients from SSF Model}
\usage{
parse_coefs(x, type)
}
\arguments{
\item{x}{SSF model object}

\item{type}{character, either RSF or Dist for model type}
}
\value{
Returns Coefficients from SSF model object
}
\description{
Extract covars for plotting from an SSF Model
}
\examples{
\donttest{coef<-as.data.frame(data.table::rbindlist(lapply(mods,parse_coefs)))}
}
\keyword{coefficients}
\keyword{competition,}
\keyword{direct}
\keyword{ssf,}
