% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefs.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Plot coefficients from SSF models}
\usage{
plot_coefs(x, plotfolder)
}
\arguments{
\item{x}{coefficients from parse_coefs}

\item{plotfolder}{folder to save plots into}
}
\value{
Returns plots of coefficients
}
\description{
Plot coefficients from SSF models
}
\examples{
\donttest{lapply(spp,plot_coefs,plotfolder='/home/puma/Desktop/DEERP/plots3/')}
}
\keyword{coefficients,}
\keyword{competition,}
\keyword{direct}
\keyword{plots}
\keyword{ssf,}
