% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_extent_to_poly.R
\name{raster_extent_to_poly}
\alias{raster_extent_to_poly}
\title{Raster extent to polygon}
\usage{
raster_extent_to_poly(x, buffer = 0)
}
\arguments{
\item{x}{raster containing extent}

\item{buffer}{amount in coordinate units to buffer the polygon}
}
\value{
spatial polygon object
}
\description{
Convert raster extent to polygon with or without buffer
}
\examples{
\donttest{dat<-raster_extent_to_poly(x = ras, buffer = 0)}
}
\keyword{ID,}
\keyword{animal}
\keyword{capture,}
\keyword{homerange}
