% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_raster.R
\name{rescale_raster}
\alias{rescale_raster}
\title{Rescale Raster 0-1}
\usage{
rescale_raster(x)
}
\arguments{
\item{x}{raster containing values to rescale}
}
\value{
rescaled raster
}
\description{
Rescale a raster from 0 to 1
}
\examples{
\donttest{dat<-rescale_raster(x = ras)}
}
\keyword{ID,}
\keyword{animal}
\keyword{capture,}
\keyword{homerange}
