% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_mod.R
\name{rf_mod}
\alias{rf_mod}
\title{RF RSF-type models}
\usage{
rf_mod(x, ras, basepath, ntree, return = FALSE)
}
\arguments{
\item{x}{output of rf_sampling}

\item{ras}{raster stack of covars}

\item{basepath}{base location to save to}

\item{ntree}{number of trees to create in RF Model}

\item{return}{logical, if TRUE, returns the predicted raster}
}
\value{
Predict spatial rf models
}
\description{
Create an RF RSF type model
}
\examples{
\donttest{system.time({ sampout<-lapply(d[1],rf_sampling,dat=df,nsamps=2,nt=5,maxnt=5,proj=dfproj,ras=nras) })}
}
\keyword{rf}
\keyword{sampling,}
