% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssf_sample.R
\name{ssf_sample}
\alias{ssf_sample}
\title{Sampling for SSF Modeling}
\usage{
ssf_sample(x, dat, ras, proj, basepath, nran)
}
\arguments{
\item{x}{list of data.frames with Month, Year and Species columns}

\item{dat}{dataframe of all data to subset}

\item{ras}{stack of rasters to extract to points}

\item{proj}{projection of point data}

\item{basepath}{base location to RF rasters}

\item{nran}{number of random points for each used}
}
\value{
Model ready DF
}
\description{
Create dataset for ssf modeling
}
\examples{
\donttest{system.time({ od<-lapply(d[1],ssf_sample, dat=df, ras=rasstack, proj=proj4string(rasstack), basepath='/home/puma/Desktop/DEERP/RF/', nran=5) })}
}
\keyword{sampling,}
\keyword{ssf}
