% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summer_sub.R
\name{summer_sub}
\alias{summer_sub}
\title{Monthly summer subset}
\usage{
summer_sub(x, spp)
}
\arguments{
\item{x}{data.frame containing relocation data}

\item{spp}{text string denoting species}
}
\value{
Returns summer subset data
}
\description{
Subset data by month/year for summer months (6, 7, 8)
}
\examples{
\donttest{dat<-summer_sub(x = dat, spp='FMD')}
}
\keyword{ID,}
\keyword{animal}
\keyword{capture,}
\keyword{homerange}
